docs: update data table demo to use playground
