+++ packages/leavittbook/src/demos/titanium-data-table/index.html
@@ -1,0 +1,58 @@
+<!DOCTYPE html>
+<head>
+  <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500&display=swap" rel="stylesheet" />
+  <link href="https://fonts.googleapis.com/css?family=Material+Icons+Outlined&display=block" rel="stylesheet" />
+  <style>
+    body {
+      font-family: Roboto, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol';
+      line-height: 1.5;
+      -webkit-font-smoothing: antialiased;
+      background-color: #fff;
+      color: #5f6368;
+
+      --app-primary-color: var(--app-link-color);
+      --app-secondary-color: #4285f4;
+      --app-text-color: #5f6368;
+      --app-light-text-color: #707175;
+      --app-dark-text-color: #202124;
+      --app-hover-color: #f9f9f9;
+      --app-border-color: #dadce0;
+      --app-menu-text-color: #5f6368;
+      --app-link-color: #1a73e8;
+      --material-primary-color: var(--app-secondary-color);
+      --material-primary-text-color: var(--app-link-color);
+      --material-error-text-color: #da3227;
+      --mdc-theme-primary: var(--app-primary-color);
+      --mdc-theme-secondary: var(--app-secondary-color);
+      --mdc-text-field-outlined-idle-border-color: var(--app-border-color);
+      --mdc-select-outlined-idle-border-color: var(--app-border-color);
+      --mdc-icon-font: 'Material Icons Outlined';
+    }
+
+    @font-face {
+      font-family: 'Metropolis';
+      font-weight: 600;
+      font-display: swap;
+      src: local('Metropolis'), url('https://www.leavitt.com/fonts/metropolis-bold.otf') format('opentype');
+    }
+
+    @font-face {
+      font-family: 'Metropolis';
+      font-weight: 500;
+      font-display: swap;
+      src: local('Metropolis'), url('https://www.leavitt.com/fonts/metropolis-semibold.otf') format('opentype');
+    }
+
+    @font-face {
+      font-family: 'Metropolis';
+      font-weight: 400;
+      font-display: swap;
+      src: local('Metropolis'), url('https://www.leavitt.com/fonts/metropolis-medium.otf') format('opentype');
+    }
+  </style>
+
+  <script type="module" src="./titanium-data-table-playground.js"></script>
+</head>
+<body>
+  <titanium-data-table-playground></titanium-data-table-playground>
+</body>
