+++ packages/leavittbook/src/demos/titanium-data-table/project.json
@@ -1,0 +1,8 @@
+{
+  "section": "Basics",
+  "files": {
+    "titanium-data-table-playground.ts": {},
+    "index.html": {}
+  },
+  "order": 0
+}
